M2k Digital: Use unsigned short values for the output of the digital part.

+++ include/libm2k/utils/channel.hpp
@@ -41,7 +41,7 @@ public:
 	bool isOutput();
 	bool isEnabled();
 	bool hasAttribute(std::string attr);
-	void write(struct iio_buffer* buffer, std::vector<short> const &data);
+	void write(struct iio_buffer* buffer, std::vector<int> const &data);
 	void write(struct iio_buffer* buffer, std::vector<double> const &data);
 	void convert(int16_t *avg, int16_t *src);
 	void convert(double *avg, int16_t *src);
