+++ src/utils/deviceout.cpp
@@ -36,7 +36,7 @@ DeviceOut::~DeviceOut()
 {
 }
 
-void DeviceOut::push(std::vector<short> const &data, unsigned int channel,
+void DeviceOut::push(std::vector<int> const &data, unsigned int channel,
 		  bool cyclic, bool multiplex)
 {
 	m_pimpl->push(data, channel, cyclic, multiplex);
